+++ src/components/MyPageScreen.vue
@@ -1,4 +1,8 @@
 <script setup>
+import MyPageBidSection from './mypage/MyPageBidSection.vue'
+import MyPageProfileCard from './mypage/MyPageProfileCard.vue'
+import MyPagePurchaseSection from './mypage/MyPagePurchaseSection.vue'
+
 const props = defineProps({
   assets: {
     type: Object,
@@ -29,83 +33,8 @@ function purchaseBadgeClass(status) {
       <h2>마이페이지</h2>
     </header>
 
-    <section class="mypage-profile-card">
-      <div class="mypage-profile-row">
-        <div class="mypage-profile-user">
-          <img :src="assets.mypageAvatar" :alt="profile.seller" class="mypage-avatar" />
-          <div class="mypage-user-copy">
-            <img :src="assets.mypageBadge" alt="" class="mypage-badge" />
-            <strong>{{ profile.seller }}</strong>
-          </div>
-        </div>
-      </div>
-
-      <div class="mypage-points-block">
-        <span>보유 포인트</span>
-        <strong>{{ profile.points }}</strong>
-      </div>
-    </section>
-
-    <section class="mypage-section">
-      <div class="mypage-section-header">
-        <h3>입찰 내역</h3>
-        <button type="button">View All</button>
-      </div>
-
-      <div class="mypage-bid-list">
-        <article
-          v-for="(item, index) in bidItems"
-          :key="`${item.title}-${index}`"
-          class="mypage-bid-card"
-        >
-          <img :src="assets.mypageProductImage" :alt="item.title" class="mypage-bid-image" />
-
-          <div class="mypage-bid-content">
-            <div class="mypage-bid-copy">
-              <span class="mypage-chip is-auction">{{ item.status }}</span>
-              <h4>{{ item.title }}</h4>
-              <div class="mypage-time">
-                <img :src="assets.clockIcon" alt="" />
-                <span>{{ item.time }}</span>
-              </div>
-            </div>
-
-            <div class="mypage-bid-prices">
-              <span>현재가</span>
-              <strong>{{ item.currentPrice }}</strong>
-              <span>내 입찰가</span>
-              <strong>{{ item.myBidPrice }}</strong>
-              <span>{{ item.date }}</span>
-            </div>
-          </div>
-        </article>
-      </div>
-    </section>
-
-    <section class="mypage-section">
-      <div class="mypage-section-header">
-        <h3>진행 중 구매 현황</h3>
-        <button type="button">View All</button>
-      </div>
-
-      <div class="mypage-purchase-list">
-        <article
-          v-for="(item, index) in purchaseItems"
-          :key="`${item.title}-${index}`"
-          class="mypage-purchase-card"
-        >
-          <div class="mypage-purchase-copy">
-            <span class="mypage-chip" :class="purchaseBadgeClass(item.status)">{{ item.status }}</span>
-            <h4>{{ item.title }}</h4>
-            <p>구매자: {{ item.buyer }}</p>
-          </div>
-
-          <div class="mypage-purchase-price">
-            <strong>{{ item.price }}</strong>
-            <span>{{ item.date }}</span>
-          </div>
-        </article>
-      </div>
-    </section>
+    <MyPageProfileCard :assets="assets" :profile="profile" />
+    <MyPageBidSection :assets="assets" :items="bidItems" />
+    <MyPagePurchaseSection :items="purchaseItems" :purchase-badge-class="purchaseBadgeClass" />
   </section>
 </template>
