+++ src/components/auction-detail/BidHistoryDrawer.vue
@@ -1,0 +1,63 @@
+<script setup>
+defineProps({
+  assets: {
+    type: Object,
+    required: true,
+  },
+  item: {
+    type: Object,
+    required: true,
+  },
+  rows: {
+    type: Array,
+    required: true,
+  },
+})
+
+defineEmits(['close', 'open-bid'])
+</script>
+
+<template>
+  <div class="detail-history-drawer-overlay" @click.self="$emit('close')">
+    <aside class="detail-history-drawer">
+      <div class="detail-history-drawer-top">
+        <button type="button" class="detail-history-drawer-close" @click="$emit('close')">
+          ×
+        </button>
+      </div>
+
+      <div class="detail-history-drawer-summary">
+        <img :src="assets.listWatchImage" :alt="item.title" class="detail-inquiry-thumb" />
+        <div class="detail-inquiry-summary-copy">
+          <strong>{{ item.title }}</strong>
+          <span>현재 입찰가</span>
+          <em>{{ item.price }}원</em>
+        </div>
+      </div>
+
+      <div class="detail-history-table">
+        <div class="detail-history-table-head">
+          <span>입찰자</span>
+          <span>입찰액</span>
+          <span>입찰 일시</span>
+        </div>
+
+        <div class="detail-history-table-body">
+          <div
+            v-for="(row, index) in rows"
+            :key="`${row.bidder}-${index}`"
+            class="detail-history-table-row"
+          >
+            <span>{{ row.bidder }}</span>
+            <strong>{{ row.amount }}</strong>
+            <span>{{ row.date }}</span>
+          </div>
+        </div>
+      </div>
+
+      <button type="button" class="detail-history-action" @click="$emit('open-bid')">
+        입찰하기
+      </button>
+    </aside>
+  </div>
+</template>
