+++ src/components/auction-detail/ReportModal.vue
@@ -1,0 +1,74 @@
+<script setup>
+defineProps({
+  assets: {
+    type: Object,
+    required: true,
+  },
+  form: {
+    type: Object,
+    required: true,
+  },
+  item: {
+    type: Object,
+    required: true,
+  },
+  reportTypes: {
+    type: Array,
+    required: true,
+  },
+})
+
+defineEmits(['close', 'submit'])
+</script>
+
+<template>
+  <div class="detail-inquiry-overlay" @click.self="$emit('close')">
+    <section class="detail-report-modal">
+      <div class="detail-inquiry-header">
+        <h3>신고하기</h3>
+        <button type="button" class="detail-inquiry-close" @click="$emit('close')">×</button>
+      </div>
+
+      <div class="detail-inquiry-summary">
+        <img :src="assets.listWatchImage" :alt="item.title" class="detail-inquiry-thumb" />
+        <div class="detail-inquiry-summary-copy">
+          <strong>{{ item.title }}</strong>
+          <span>현재 입찰가</span>
+          <em>{{ item.price }}원</em>
+        </div>
+      </div>
+
+      <div class="detail-report-fields">
+        <div class="detail-report-field">
+          <span>신고 유형 <em>*</em></span>
+          <div class="detail-report-options">
+            <button
+              v-for="type in reportTypes"
+              :key="type"
+              type="button"
+              class="detail-report-option"
+              :class="{ 'is-selected': form.type === type }"
+              @click="form.type = type"
+            >
+              <span class="detail-report-radio" />
+              <span>{{ type }}</span>
+            </button>
+          </div>
+        </div>
+
+        <label class="detail-report-field">
+          <span>신고 내용 <em>*</em></span>
+          <textarea
+            v-model="form.detail"
+            class="detail-report-textarea"
+            placeholder="신고 사유에 대한 구체적인 내용을 작성해주세요."
+          />
+        </label>
+      </div>
+
+      <button type="button" class="detail-report-submit" @click="$emit('submit')">
+        신고하기
+      </button>
+    </section>
+  </div>
+</template>
