+++ src/components/inspection/InspectionShippingModal.vue
@@ -1,0 +1,48 @@
+<script setup>
+defineProps({
+  form: {
+    type: Object,
+    required: true,
+  },
+})
+
+defineEmits(['close', 'submit'])
+</script>
+
+<template>
+  <div class="inspection-shipping-overlay" @click.self="$emit('close')">
+    <section class="inspection-shipping-modal">
+      <div class="inspection-shipping-header">
+        <h3>배송 정보 등록</h3>
+        <button type="button" class="inspection-shipping-close" @click="$emit('close')">×</button>
+      </div>
+
+      <div class="inspection-shipping-form">
+        <label class="inspection-shipping-field is-small">
+          <span>택배사 <em>*</em></span>
+          <div class="register-select-wrap">
+            <select v-model="form.company">
+              <option value="" disabled>선택</option>
+              <option value="우체국 택배">우체국 택배</option>
+              <option value="CJ대한통운">CJ대한통운</option>
+              <option value="한진택배">한진택배</option>
+            </select>
+          </div>
+        </label>
+
+        <label class="inspection-shipping-field">
+          <span>송장 번호 <em>*</em></span>
+          <input
+            v-model="form.invoiceNumber"
+            type="text"
+            placeholder="송장 번호를 입력해 주세요."
+          />
+        </label>
+      </div>
+
+      <button type="button" class="inspection-shipping-submit" @click="$emit('submit')">
+        등록하기
+      </button>
+    </section>
+  </div>
+</template>
